requirements.txt
discord.py
python-dotenv~=1.1.1
dropbox~=12.0.2
boto3~=1.39.4